refactor(davinci): split v-model check out of the 350-line pass file
P2-10's VueCssBind exhaustiveness arm crossed the source-length gate.
